python-backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
distro==1.9.0
dnspython==2.7.0
email_validator==2.2.0
exceptiongroup==1.3.0
fastapi==0.111.0
fastapi-cli==0.0.7
filelock==3.18.0
fsspec==2025.5.1
future==1.0.0
groq==0.28.0
h11==0.16.0
hf-xet==1.1.4
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.33.0
idna==3.10
Jinja2==3.1.6
lxml==5.4.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
numpy==2.2.6
openai==1.6.1
orjson==3.10.18
packaging==25.0
pydantic==2.6.4
pydantic_core==2.16.3
Pygments==2.19.1
pymavlink==2.4.47
python-dotenv==1.0.0
python-multipart==0.0.7
PyYAML==6.0.2
regex==2024.11.6
requests==2.31.0
rich==14.0.0
rich-toolkit==0.14.7
safetensors==0.5.3
shellingham==1.5.4
sniffio==1.3.1
starlette==0.37.2
tokenizers==0.15.2
tqdm==4.67.1
transformers==4.36.0
typer==0.16.0
typing_extensions==4.14.0
ujson==5.10.0
urllib3==2.5.0
uvicorn==0.27.1
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
